requirements.txt
requests~=2.32.4
allure-python-commons~=2.14.3
allure-pytest~=2.15.0
pytest~=8.4.1